golf: semi-working golf hole

+++ toontown/golf/GolfHoleBase.py
@@ -208,7 +208,7 @@ class GolfHoleBase:
         frameCount = 0
         lift = 0
         startTime = GolfGlobals.BALL_CONTACT_FRAME / 24
-        startFrame = startTime * self.FPS
+        startFrame = int(startTime * self.FPS)
         for frame in range(startFrame):
             self.simulate()
             self.setTimeIntoCycle(self.swingTime + float(frameCount) * self.DTAStep)
